% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_coralmz.R
\docType{data}
\name{coralmz}
\alias{coralmz}
\title{coralmz}
\format{A tibble with 230 rows and 7 variables:
\describe{
  \item{reported_mz}{the m/z reported}
  \item{reported ion}{the ion reported}
  \item{monoiso_mass}{the monoisotopic mass}
  \item{molec_formula}{the molecular formula}
  \item{reported_name}{the reported name}
  \item{taxon}{the taxonomic information}
  \item{reference}{the reference where the information was reported}
}}
\source{
Benjamin R. Gordon
}
\usage{
coralmz
}
\description{
A dataset of known coral metabolites identified using mass spectrometry.
}
\keyword{datasets}
