% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_coralmz.R
\name{write_coralmz}
\alias{write_coralmz}
\title{write coralmz}
\usage{
write_coralmz(save.rda = FALSE)
}
\arguments{
\item{save.rda}{Logical indicating if an .rda file should be saved to
\code{/data}.}
}
\value{
Returns a dataframe of class \code{tbl_df}
}
\description{
Converts \code{./data/coralmz.csv} to an \code{.rda} file
}
\details{
\code{write_coralmz()} loads \code{coralmz.csv}, cleans up the
headers and saves \code{coralmz.rda} to the \code{./data} directory.
}
\author{
Benjamin R. Gordon
}
